requirements.txt
blis==0.4.1
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
cymem==2.0.3
Flask==1.1.2
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
streamlit
streamlit_bokeh
MarkupSafe==1.1.1
murmurhash==1.0.2
numpy==1.18.2
plac==1.1.3
preshed==3.0.2
requests==2.23.0
six==1.14.0
spacy==2.2.4
srsly==1.0.2
thinc==7.4.0
https://download.pytorch.org/whl/cpu/torch-1.4.0%2Bcpu-cp37-cp37m-linux_x86_64.whl
torchtext==0.3.1
tqdm==4.45.0
urllib3==1.25.8
wasabi==0.6.0
Werkzeug==1.0.1
zipp==3.1.0
